refactor(Configurations.interfaces): renamed from Configurations.devices
renamed ...
