+++ lib/v1/configurations/base.js
@@ -277,7 +277,7 @@ module.exports = {
   level: require('./items/level'),
   taxes: require('./items/taxes'),
   analytics: require('./items/analytics'),
-  devices: require('./items/devices'),
+  interfaces: require('./items/interfaces'),
   custom_dashboards: {
     oneOf: [
       {
