+++ lib/v1/configurations/items/interfaces.js
